test(editor): split the EditorLogicTests monolith by topic
Audit #250 F050. The single 2,033-line file held some twenty unrelated
suites and was the merge-conflict magnet the audit measured (59 commits
in three months; it grew another 200+ lines during this campaign
alone). The tests now live in seven topic files plus the existing
InfrastructureTests, following the convention AnalysisResponseTests
established; EditorLogicTests.cpp keeps only the harness, the
expect/require wrappers and main() (171 lines), and the run order is
unchanged.

Bodies moved verbatim (verified character-for-character against HEAD;
787 assertion-call tokens before and after). The suite passes with
exactly the same count as before the split: 2,917 checks.
