+++ Tests/EditorLogicTests/EditorLogicTestSupport.h
@@ -17,6 +17,25 @@ void expectEqual(const QStringList& actual, const QStringList& expected, const Q
 void requireTrue(bool value, const QString& message);
 void requireEqual(int actual, int expected, const QString& message);
 
+void testConvolutionPathHelper();
+void testAnalysisWorkerRecovery();
+void testUpdateInfoFormatter();
+void testVelopackUpdateInfo();
+void testVelopackGitHubRelease();
+void testVelopackFeeds();
+void testFilterCardDescriptors();
+void testSubwooferRoutingDescriptors();
+void testSubwooferRoutingCrossoverRecipes();
+void testFilterCardDepths();
+void testFilterCardBuildPlans();
+void testConfigImport();
+void testLegacyMigrationScanAndPolicy();
+void testChannelSelectionModel();
+void testDeviceSelectionModel();
+void testMultiConvolutionRoutingAdapter();
+void testStageSelectionModel();
+void testStudioRoutingModel();
+void testRoutingFold();
 void testAnalysisResponseBinArithmetic();
 void testAnalysisResponseEmptyAndLatency();
 void testResponseCurveMatchesTheLegacyMagnitudePath();
@@ -35,6 +54,7 @@ void testConfigFileCodec();
 void testConfigFileCodecRetriesAtomicReplaceAfterReaderCloses();
 void testConfigFileCodecPreservesExistingFileWhenAtomicReplaceFails();
 void testConfigFileCodecRejectsPartialRead();
+void testMemoryHelperConstructReleasesStorageWhenConstructorThrows();
 void testOwnedBackgroundTaskJoinsAndStartsOnlyOnce();
 void testUpdateElevationPolicyUsesOnePromptForEditorUpdates();
 void testTheSkinRosterIsTheOneList();
@@ -44,6 +64,8 @@ void testEverySkinSheetResolvesAllThemeTokens();
 void testEditableValueTextUsesDisplayedDecimalFormatFirst();
 void testBenchmarkBatchPlanUsesOnlyComparableFullBatches();
 void testFileReferenceControllerOwnsPathState();
+void testFilterListModel();
+void testFilterListUndo();
 void testFilterCommandCatalogRoster();
 void testFilterCommandCatalogIconsExistOnDisk();
 void testFilterCommandCatalogTemplateRoster();
